Initial commit and implementation for a simple timing harness
Named hawking. Pass it the command and options you want to run
and it will run them and then output the last thing written to
stdout by the command. It also outputs the exit code, and the time
that the command ran for in nanoseconds, milliseconds and seconds.

Written for use in hacknightslc.com's prime number drag races.